+++ compiler/src/printLinear.mli
@@ -1,3 +1,4 @@
+val pp_wsize : Format.formatter -> Wsize.wsize -> unit
 val pp_prog :
   ('reg, 'regx, 'xreg, 'rflag, 'cond, 'asm_op, 'extra_op) Arch_extra.extended_op Sopn.asmOp ->
   Conv.coq_tbl -> Format.formatter ->