+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 # [unreleased]
 
+## New features
+
+- x86 intrinsics that accept a size suffix (e.g., `_128`) also accept, with a
+  warning, a vector suffix (e.g., `_4u32`)
+  ([PR #303](https://github.com/jasmin-lang/jasmin/pull/303)).
+
 ## Other changes
 
 - Instruction selection for `x86_64` recognizes shifts (rotations, etc.) by